commit API de ejemplo porque parece que no sube

+++ src/routes/user.js
@@ -2,39 +2,42 @@ const { request, response, application } = require("express");
 const express = require("express");
 const user = require("../models/user");
 const router = express.Router();
-const users = require("../middleware/userFunc")
-const path = require('path');
-const storage = require('../middleware/storage')
+const users = require("../middleware/userFunc");
+const path = require("path");
+const storage = require("../middleware/storage");
+const token = require("../middleware/authTokens");
 const app = express();
-const multer  = require('multer');
+const multer = require("multer");
 //________________Register__________________
 
 router.post("/register", async (request, response) => {
   try {
-    const newUser = await users.create(request)    
+    const newUser = await users.create(request);
     const userSaved = await newUser.save();
-    await users.register(newUser,response)
+    await users.register(newUser, response);
   } catch (error) {
-    response.json({message:error});
+    response.json({ message: error });
   }
 });
 
 //_____________________Login_____________________
 
 router.post("/login", async (request, response) => {
-    await users.login(request,response)
+  await users.login(request, response);
 });
 
 //_____________________gets__________________________
 router.get("/", async (request, response) => {
   try {
-    const verified = token.verify(request.headers['authorization'])
-  if(verified){
-    const users = await user.find();
-    response.json({status:"200",data:users});
-  }else{
-    response.json({status:"403",message:"forbidden"})
-  }
+    const verified = token.verify(
+      request.headers["authorization"].split(" ")[1]
+    );
+    if (verified) {
+      const users = await user.find();
+      response.json({ status: "200", data: users });
+    } else {
+      response.status(403);
+    }
   } catch (error) {
     response.json(error);
   }
@@ -43,8 +46,15 @@ router.get("/", async (request, response) => {
 router.get("/:id", async (request, response) => {
   const { id } = request.params;
   try {
-    const userFinded = await user.findById(id);
-    response.json({status:"ok",data:userFinded});
+    const verified = token.verify(
+      request.headers["authorization"].split(" ")[1]
+    );
+    if (verified) {
+      const userFinded = await user.findById(id);
+      response.json({ status: "ok", data: userFinded });
+    } else {
+      response.status(403);
+    }
   } catch (error) {
     response.json(error);
   }
@@ -52,60 +62,84 @@ router.get("/:id", async (request, response) => {
 
 //____________________________Updates______________________
 router.put("/:id", async (request, response) => {
-  await users.update(request,response)
+  const verified = token.verify(request.headers["authorization"].split(" ")[1]);
+  if (verified) {
+    await users.update(request, response);
+  } else {
+    response.status(403);
+  }
 });
 
 router.put("/adm/:id", async (request, response) => {
-  await users.admUpdate(request,response)
+  const verified = token.verify(request.headers["authorization"].split(" ")[1]);
+  if (verified) {
+    await users.admUpdate(request, response);
+  } else {
+    response.json(403);
+  }
 });
 
 //_____________________delete________________
 router.delete("/:id", async (request, response) => {
   const { id } = request.params;
   try {
-    const removed = await user.deleteOne({ _id: id });
-    response.json({status:"200",data:removed});
+    const verified = token.verify(
+      request.headers["authorization"].split(" ")[1]
+    );
+    if (verified) {
+      const removed = await user.deleteOne({ _id: id });
+      response.json({ status: "200", data: removed });
+    } else {
+      response.status(403);
+    }
   } catch (error) {
-    response.json({error:error});
+    response.json({ error: error });
   }
 });
 
-
 //__________________GET-img___________________
 
-router.use('/img',express.static(path.join(__dirname,'../../img/user')));
+router.use("/img", express.static(path.join(__dirname, "../../img/user")));
 
-app.get('/img/:image', (req, res) => {
-    res.sendFile(path.join(__dirname, 'img', req.params.image));
-  });
+app.get("/img/:image", (req, res) => {
+  const verified = token.verify(request.headers["authorization"].split(" ")[1]);
+  if (verified) {
+    res.sendFile(path.join(__dirname, "img", req.params.image));
+  } else {
+    response.status(403);
+  }
+});
 
 //_____________________Post-img___________________________
-  const upload = multer({storage: storage.storageUser})
+const upload = multer({ storage: storage.storageUser });
 
-  router.post('/img', upload.single('image'), async (req, res) => {
-    res.status(201).send('Image uploaded succesfully')
-  })
+router.post("/img", upload.single("image"), async (req, res) => {
+  res.status(201).send("Image uploaded succesfully");
+});
 
-
-  router.put('/deleteimg/:id', async (req, res) => {
-    try {
+router.put("/deleteimg/:id", async (req, res) => {
+  try {
+    const verified = token.verify(req.headers["authorization"].split(" ")[1]);
+    if (verified) {
       const { id } = req.params;
       picture = {
-        picture:"example-user.png"
+        picture: "example-user.png",
       };
-   
+
       const updated = await user.updateOne(
         { _id: id },
         {
-          $set:picture
+          $set: picture,
         }
       );
-      res.json({status:"200",post:updated});
-    } catch (error) {
-      res.json({status:"500",message:error});
-      console.log(error)
+      res.json({ status: "200", post: updated });
+    } else {
+      res.status(403);
     }
-  })
-
+  } catch (error) {
+    res.json({ status: "500", message: error });
+    console.log(error);
+  }
+});
 
 module.exports = router;
